Move Val compiler from test_expr_simplifier.cpp to own file
